fix diffusers schedulers
